Terminate agent if it tries to register to unregistered zone
When aircrew use specific zone using --zone option, it start
listening to that zone, but it sends the registration request
over control channel. ControlTower receives the request and if
it doesn't find the peer zone, it sends and error message on
the channel that agent specified.
Agent receives the message and terminate.

Signed-off-by: Anil Kumar Vishnoi <avishnoi@redhat.com>