Add error handling for zones.
In currently implementation, if agent tries to subscribe to
a zone that is not registered with the controltower, controltower will
send an error message. Given that when aircrew sends a peer registration
request, it will subscribe to the channel (same name as zone)
and controltower will use that channel to send the error message.
Once aircrew agent receives the error message, it terminates.
This approach currently doesn't broadcast on the controller
channel, but using the channel that agent automatically
subscribes to duing bootup.

This behavior might change as we evolve the control flow in future.

This PR also includes some initial refactoring of the aircrew.

Signed-off-by: Anil Kumar Vishnoi <avishnoi@redhat.com>