Determine public IP address using stun (#135)
* Determine public IP address using stun

Change to using STUN for determining the public IP address when the
--public-address option is used. This patch just hard codes the STUN
server used (google). Someone may complain and request that it be
configurable at some point, but I'd rather just skip adding another
config option for the moment.

There are more go.mod changes than expected here. It's the result of
running "go mod tidy" after adding the new dependency.

Closes issue #31

Signed-off-by: Russell Bryant <rbryant@redhat.com>

* apex: rename --public-network option to --stun

This patch renames the --public-network option to --stun. The new
option name is explicit about what is being turned on. I also just
find it less confusing, because you would normally use this when on a
private network, not a public one.

Signed-off-by: Russell Bryant <rbryant@redhat.com>