Fix tags for apex images
Currently CI image job is pushing the image under main tag.

So everytime you run docker-compose locally it builds the
images because it doesn't find latest tag in quay.io. If
you want to pull the images from quay.io, you need to change
the tag to main.

CI endup building the images (which is required) for the same
reason.

This patch, publishes the images under latest tag, and explictly
asking CI to build the images.

Signed-off-by: Anil Kumar Vishnoi <avishnoi@redhat.com>