Stash usernames in our DB and show them in the UI
This patch stashes the UserName in the request context in the auth
middleware, and then users it when creating a user in our database.
The UI is also updated to show the username in the User list instead
of IDs. The ID is still visible in the user detail view.

If a UserName changes, we'll detect it and update our cache of users.
This is still just a simple cache and the UserName is subject to going
stale if it changes in the auth system, but doesn't log in to apex.

Signed-off-by: Russell Bryant <rbryant@redhat.com>