ui: List zones by name in the Device listing
On the Device list page, one column was the list of Peer UUIDs
associated with this Device. Instead of calling the column "Peers," it
is now called "Peered Zones," and the list is the names of the zones
the device is peered with.

In passing, I dropped the Public Key from the Device List, because it
didn't seem that interesting. It's available in the Device detail
page.

One slight downside of the change is that the <ChipField /> used
previously looked nicer than just hyperlinked text, but I couldn't
figure out how to still use that UI element. Overall I think this is
an improvement, though.

Signed-off-by: Russell Bryant <rbryant@redhat.com>