README.md: Add Guiding Principles
This change introduces a set of guiding principles for the project.
These principles should be things that the whole team agrees on and
feels passionately about. These principles should be used to guide
decision making as the project continues to evolve.

There's enough to say about security that we point to a new, dedicated
document where we can discuss all things security for Apex. There's
more that could be added here in the future, like details on key
management.

There are a few other minor cleanups to other text included, as well.

Signed-off-by: Russell Bryant <rbryant@redhat.com>