HTTPS ALL THE THINGS!
Adds HTTPS config to the dev overlay via a Kustomize Component.
Start adding the scaffold for an operate-first overlay - we'll need to
figure out how to use cert-manger there.
Updates dev README on how to add the self-signed CA to the root store.
This will push us to update the deployment on K8s vs. up/down of cluster
to avoid having to re-trust the cert, but that seems a fair tradeoff.

Minikube will break as it will need to patch the ingress controller to
add the necessary flag for SSL passthrough.

In staging/production this shouldn't be needed since the auth server
will be reachable via HTTPS and have a valid certificate.

Also moves tests to testcontainers/testcontainers-go from dockertest.
This allows for tests to be run using rootful podman (because we still
need privileges).

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>