ui: Improve title to PeerShow page (#207)
The previous title for the Peer Show page was "Peer #<id>". The titles
for all other pages was updated to a more friendly representation by
changing the recordRepresentation on the Resource. I wasn't able to get
that to work for peers.

Instead, this just explicitly updates the title for the page showing
details of a single peer. Now the title shows "Peer: Device <device>
to Zone <zone>."

There are no other places at the moment that would benefit from
updating the recordRepresentation for peers, so this seems good
enough.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

Signed-off-by: Russell Bryant <rbryant@redhat.com>