apiserver: Introduce feature flags
This commit introduces feature flags to the apiserver. It adds the
first flag, multi-zone, which enables/disables multi-zone support.
This flag is currently set to off for all users.

The flags are also exposed via an API using the
go-feature-flag-relay-proxy.

My intention is that the flags are stored in git and retrieved via
https. This is an easy way to let go-feature-flags periodically
refresh the feature flags in the background. Right now the flags are
loaded from a file in the apiserver because of an open issue with CA
trust.

Signed-off-by: Russell Bryant <rbryant@redhat.com>