Makefile: remove apiserver from all target (#222)
A recent change dropped the apiserver targets from the Makefile. The
'all' target still had an 'apiserver' dependency, but no 'apiserver'
target existed. We exclusively run the apiserver in a container, so we
don't need to build it directly here.

Re-add APISERVER_DEPS, as it's used as a dependency for the go-lint
target, and that still makes sense.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

Signed-off-by: Russell Bryant <rbryant@redhat.com>