docs: Remove outdated docs (#238)
* hack: Remove hack/README.md

This document is completely out of date and is replaced by some
content in doc/README.md.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

* docs: Remove outdated docs

A lot of the content here has gone stale. The diagrams reflected an
old architecture. Instructions still referenced docker-compose. Other
details have changed.

I believe it's better to have no docs than incorrect docs, so just
remove these until someone has time to re-write something similar.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

* README.md: Shorten the disclaimer

Saying that the project is experimental is enough.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

* docs: Update interactive enrollment sample

The text here was a bit outdated. Update it to reflect the current
state of this workflow.

Signed-off-by: Russell Bryant <rbryant@redhat.com>