Control multiple zone support with a feature flag
This commit introduces the first phase of feature flags and uses a
feature flag to disable support for multiple zones (multi-zone).

The feature-flag backend is represented by the FFlags API, which, for
now is just a simple backend with defaults that can be changed with
an environment variable. Eventually, we can replace this with one of
several other options that will allow us to do more advanced feature
controls.

Feature flags are exposed via our own API. They may be listed, or a
specific flag may be queried.

The UI consumes the API to know whether multiple zones are supported.
If they are not supported, the "Create Zone" button is disabled (or
greyed out).

The Zone create/delete APIs also check the feature flags in the API
server. These methods will be denied if the flag has multi-zone
support off.  Unit tests exercise the zone create/delete methods with
support on and off.

The dev deployment runs with multi-zone support on by default, but the
prod deployment has it off.

Closes #213
Closes #214

Signed-off-by: Russell Bryant <rbryant@redhat.com>