deploy: Use a single postgres cluster
Since the apiserver and ipamdb have a shared fate we may as well use one
postgres instance - we can always add more replicas to scale it out.
If the apiserverdb was deleted but not ipamdb, we would need to force
all leases to be released.
Similarly, if the ipamdb was deleted but not the apiserverdb, we could
issue leases for overlapping CIDRs as leases for allocated CIDRs were
not re-acquired.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>