docs: Document the use of --child-prefix (#269)
* docs: Document the use of --child-prefix

Add a "Additional Feature" section to the docs, and document how the
Apex agent can be instructed to act as a Subnet Router.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

I previously stripped a bunch of docs because they included diagrams
and instructions that were out of date. This scenario was part of it
and it shows an example of using child-prefix. I dropped the
out-of-date diagram and a few other minor bits and have restored it
here.

Signed-off-by: Russell Bryant <rbryant@redhat.com>