apex: Add unix socket control interface
Add a unix socket control interface on linux. This interface uses
jsonrpc and supports checking the status apexd. It will report back
the info needed to register the device if it's waiting for auth. This
would be one way to get info when running apex as a daemon. The other
would be from the daemon logs.

apexctl now knows how to retrieve this status information by running:

    apexctl apexd status

We now embed a version at build time into each binary. This is done
using ldflags to set the value of main.Version.

apexctl can report its own version.

    apexctl version

apexctl can also report the version of a running apexd.

    apexctl apexd version

When running subcommands of "apexctl apexd" other than "version",
apexctl will first validate that the apexctl and apexd versions match.

Signed-off-by: Russell Bryant <rbryant@redhat.com>