Fix Token Expiry Issues
@nerdalert noted that tokens created with `--with-token` were not being
properly renewed since the `Expiry` field in the `oauth2.Token` was nil.
This is because we reconstructed those tokens from JSON, which isn't a
supported operation in the oauth2 library.

To fix this, we let apex request the token on behalf of the user if it's
provided with a --username and --password

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>