apexd: Refactor buildPeerConfig()
I started looking at this when trying the gocyclo linter and it
pointed out this method as one of the most complex. I was able to
simplify it a bit by:

 - only checking if the peer in the large for loop is ourself in one
   place instead of duplicating it throughout the rest of the loop.

 - combining some conditional checks instead of having multiple levels
   of nested if statements.

I'm not introducing the gocyclo linter in the tree yet, as the
required refactorings would disrupt some other work people have in
flight.

Signed-off-by: Russell Bryant <rbryant@redhat.com>