docs: Document our design process
We have a `docs/design/` directory already with a bit of content in
it. This PR incldues a `README.md` for this directory with an overview
of our lightweight design process. It also introduces a design
proposal template that lists the essential information that should be
included. I expect the template to evolve over time to reflect things
we find most improtant to address for this project.

Signed-off-by: Russell Bryant <rbryant@redhat.com>