clean up: reduce code duplication in stun.go, use a dynamic port the wg interface that we can open.
Using a random high port for the wg is not guaranteed to not be in use by another process, so lets ask the OS for
a dynamic port instead.

This now avoids an additional stun request once we have determined that the device is behind a symmetric nat.

Wait for apex to be deployed when running `make recreate-db`

Signed-off-by: Hiram Chirino <hiram@hiramchirino.com>