design: Add a project name proposal.
The project's github repo should move. We will need a new github
organization to host the project. The `apex` org is taken, so one
option is to create a github org with some variable of `apex` like
`apex-mesh` or `apex-hosted`. The other option is to rename the
whole project to something more unique.

This PR proposes a full project rename.

Issue #428

Signed-off-by: Russell Bryant <rbryant@redhat.com>
Co-authored-by: Anil Kumar Vishnoi <avishnoi@redhat.com>
Co-authored-by: Dave Tucker <dave@dtucker.co.uk>