deploy: Check for new commits to main
The deploy workflow fails often if additional PRs have been merged by
the time this job runs. This change will check to see if there are any
commits in main that are not in the branch the job is running on. If
so, it will not attempt to push a commit that would fail.

Signed-off-by: Russell Bryant <rbryant@redhat.com>