docs: Reorganize top-level docs
`docs/README.md` was getting pretty long. I split it up into a few
pieces: deploying the nexodus service, a nexodus user guide, and a
nexodus development guide.

In passing, I updated the Concepts section to reflect changes made in
recent weeks.

Lastly, I dropped the reference to `solution-architecture.md`, as it
is still a little bit out-of-date with the name change and details on
how things operate. It's still in tree, just not linked from the main
docs page.

Signed-off-by: Russell Bryant <rbryant@redhat.com>