gen-docs: Add job pre-reqs and pretty Makefile output
Add pretty output for running `make gen-docs`. Full output from the
tool can be seen when you run with `NOISY_BUILD=y` like this:

    $ NOISY_BUILD=y make gen-docs

Also make some other jobs depend on the gen-docs job. It runs in 20%
of the time of go-lint, so make jobs that depend on go-lint depend on
this one too. If the docs are broken, don't waste time running all the
more consuming jobs.

Signed-off-by: Russell Bryant <rbryant@redhat.com>