docs: Add a style guide and related updates
Add a style guide for Nexodus specific guidelines followed in our
documentation. The first is how to refer to Nexodus as a project or as
a specific binary.

In passing I fixed a few spots that used `nexodus` as a command
instead of `nexd`. It was probably broken by accident in the name
converstion.

Signed-off-by: Russell Bryant <rbryant@redhat.com>