nexd: Add an L4 proxy mode
This commit makes use of the ability to run entirely in userspace mode
to implement basic L4 proxy functionality that requires no elevated
privileges.

This mode can be used in a Kubernetes Pod where `nexd` is not able to
create a new network device. It's possible to grant a pod the access
it needs, but in many cases an application developer does not have
this level of access available.

Multiple ingress or egress proxy rules may be specified at the same
time, but they all must be specified as command line arguments.
Runtime management of rules will come later.

Documentation on how to use this mode can be found in
`docs/user-guide/nexd-proxy.md`.

Signed-off-by: Russell Bryant <rbryant@redhat.com>