design: Add Nexodus network to our project terminology
Standardize on how we refer to a network managed by Nexodus in
documentation. Earlier in the project we referred to this as a
"zone," but that construct is gone for now. At some point a network
slicing construct may be reintroduced within organizations, but for
now just use the more generic "network" terminology.

Signed-off-by: Russell Bryant <rbryant@redhat.com>