nexd: Move run code into its own function
This moves the run code into its own function that can be
paramterized. This will be used when adding subcommands that alter the
behavior.

Also remove some code that was a placeholder and never implemented.
This cleanup probably belongs elsewhere anyway. It's going to be
dependent on the mode nexd is running in, the OS, and eventually the
tunneling mechanism.

Signed-off-by: Russell Bryant <rbryant@redhat.com>