e2e: Add nexd proxy e2e tests
Add a test for `nexd proxy` ingress and egress rules. Both tests have
two nodes, one running normal nexd and the other running nexd proxy. One
side has an http server and the other runs curl. The roles switch based
on which scenario is under test.

I added python3 and curl to the ubuntu test container for this scenario.

There's one hack in a test with a TODO comment to improve. It makes an
assumption that our ipam is handing out IPs sequentially. We can't get
the IP like we do normally because we can't read wg0 from the container
since there's no interface. We could either add a new `nexctl nexd`
command to get wireguard state, or we could query the nexodus service
for the required device info. The hack at least works for now.

Signed-off-by: Russell Bryant <russell.bryant@gmail.com>