Separate the agent and relay function.
This commit includes the following changes:
*.* Refactor the common agent field into its own struct,
so it can be used by both relay and agent logic.
Also it makes nexctl's life easy to talk to both relay and agent.

*.* Separate the Agent and Relay in its own structs and
use wireguard, apiserver and nexodus structs as nested struct,
 with fields (e.g --child-prefix only belongs to agent) specific
to the implementation. This simplifies the conditions,
minimizes duplicate code and improves the readability. 

Signed-off-by: Anil Kumar Vishnoi <avishnoi@redhat.com>