Add router and relay command
Router command separate the --child-prefix option
from the normal agent option. That helps userspace
wireguard implementaiton to not throw the not-supported
error.

Relay command is added to start the agent with relay
capability. --discovery-node flag is wrapped under the
relay command as a flag --enable-discovery. Discovery
node is added as a flag, rahter than a separate command
is because we would like to run both the function on the
same node, and you can't fire two commands in single cli
command.

Signed-off-by: Anil Kumar Vishnoi <avishnoi@redhat.com>