ci: Add a periodic job for govulncheck
Run a tool that checks the Go vulnerability database for CVEs that
affect our code. Dependabot checks for CVEs, too, but I thought we
could turn this on for a while to see if it helps find anything else.

Find more info on the tool here: https://go.dev/blog/vuln

Signed-off-by: Russell Bryant <rbryant@redhat.com>