Exit the nexd agent in the scenarios where the agent node itself doesn't show up in the peer list.
There are possible scenarios where we would like node
agent to exit itself, and not attempt to auto onboard:
*) when the apiserver db is wiped
*) when user manually delete the device from org using
either gui or api.

In both the above scenario, the device itself won't
show in the peer listing of it's org, so this condition
can be used to exit the nexd and avoid onboarding based
on the stored apitoken.

Signed-off-by: Anil Kumar Vishnoi <avishnoi@redhat.com>