e2e: Add TestProxyEgressUDP
This test sets up a single egress UDP proxy rule. It ensures that a
UDP packet makes it from a client, through the proxy, and to a server
on another Nexodus device. It also ensures that a reply packet sent
from that UDP server makes it back to the original client.

Signed-off-by: Russell Bryant <rbryant@redhat.com>