nexd: Open a unix socket in proxy mode
Turn on the unix socket interface for proxy mode. Instead of using
`/run/nexd.sock`, it defaults to `nexd.sock`. This is a safer default
as a place we can write to. The primary use case for this mode is a
container, so it seems fine to me.

Update `nexctl` to account for this possible path, as well.

Signed-off-by: Russell Bryant <rbryant@redhat.com>