nexd: Set default service URL
If nexd is run without a service URL, use a default URL that is set at
build time. If no `NEXODUS_BUILD_PROFILE` is set, it will use the `dev`
profile which will use the local dev service URL. The build profile can
be set to `qa` or `prod` which will set `qa.nexodus.io` or
`try.nexodus.io` as the default URLs.

For the `nexd` container image that gets published to quay.io and the
binary builds published to s3, they are now build with the `prod`
profile in CI.

For those using the rpm, update the default in the sysconfig file
to our public service. This would be the most likely config for an
install using the package.

This change also updates `nexctl` with the same behavior.

A side effect of this change is that `nexd` and `nexctl` in the
container image now have a version string set properly.

Signed-off-by: Russell Bryant <rbryant@redhat.com>