s/controller/service/
Long ago we renamed the controller to the apiserver, so there's
nothing named "controller" anymore. Update user facing docs and
manifests to use "Nexodus Service", as proposed in the
`project-name.md` document.

I fixed some spelling and grammar issues along the way. There are
still many controller references in the code and in ansible playbooks,
but I left those alone for now. I felt this was good enough to call
this issue complete.

Closes #224

Signed-off-by: Russell Bryant <rbryant@redhat.com>