deploy: Create nexodus-monitoring kustomization
observability-operators are installed into the observability namespace,
but are cluster-scoped so will work on all namespaces. This step is only
required for local development of the stack.

The 'monitoring' component of nexodus has been factored out into a
dedicated 'nexodus-monitoring' stack. This can be deployed locally, but
using the `rosa` overlay can be customized for deployment into our
OpenShift install - and can be set to monitor both the Prod and QA
deployments.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>