nexd: Make temporary dns failures a debug log instead of error
When nexd has a temporary failure to resolve DNS for the api server,
just log it to debug instead of error. This is a normal, expected
condition to hit periodically. In my case, I run `nexd` on my laptop,
and I get this message a lot in my logs when moving aronud, suspending
and resuming, and connecting to different networks.

Signed-off-by: Russell Bryant <rbryant@redhat.com>