nexd: Reject proxy rules with a blank dest host
Proxy rules with a blank destination host should be considered invalid.
Add a check for this and a test case.

I have observed a failure in e2e tests where proxy rules were configured
this way as a result of an earlier error. The original error would have
been easier to spot if the rule had been outright rejected.

Signed-off-by: Russell Bryant <rbryant@redhat.com>