e2e: Add logging to TestProxyInvalidConfig
I'm seeing failures in this test in CI that I have not been able to
reproduce locally. In the failure case, a test case will return without
an error. Add logging of any output returned in case that gives any
hints.

I also refactored the test to DRY it up.

Signed-off-by: Russell Bryant <rbryant@redhat.com>