e2e: Fix major flaw in TestProxyInvalidConfig
It turns out this test was never actually testing what it was supposed
to. The intent was to start `nexd` with invalid proxy rules and verify
that it exited with an error. The problem is that a very critical
argument was left out ... `nexd`. What was happening in every test case
was that it tried to run a command called `--username`. Of course, this
fails. Every so often, containerExec() would not return an error, but
since I added debug output, I was able to spot the error message that
clued me in to what went wrong.

```
    proxy_test.go:619: nexd output: �OCI runtime exec failed: exec
failed: unable to start container process: exec: "--username":
executable file not found in $PATH: unknown
 ```

Big oops.

This change fixes it so the test actually runs `nexd`. It also stops
printing the debug output every time and only dumps it if the command
unexpectedly returns success.

Closes #960

Signed-off-by: Russell Bryant <rbryant@redhat.com>