apiserver: implement ext_autz grpc endpoint
We were stuffing the AccessToken in a browser cookie. With this
change we get the AccessToken from the redis session.  This lowers
the amount of data placed into browser cookies to just being the
Redis session id.

This will also give us a place to implement per device API token in
the future.

Signed-off-by: Hiram Chirino <hiram@hiramchirino.com>