rpm: Restart nexodus on package upgrade
I noticed that nexodus was not getting restarted as I regularly updated
the package. The reason was that I had used `%systemd_postun` instead
of `%systemd_postun_with_restart`. I went ahead and removed the extra
stuff checking to make sure these systemd scriptlets exist, because they
seem to exist for both distro versions we are building rpms for right
now. This simplifies the spec file a bit.

I also dropped the build requirement on `make`. It is not used.

Signed-off-by: Russell Bryant <rbryant@redhat.com>