hack: Add tmux to nexd container
This change includes a couple of updates to `make run-nexd-container`.

First, the commands for running `nexd` against a dev instance, qa, or
prod, are now in the bash history. You can hit the up arrow to choose
which one you want to execute.

Start up with tmux. You can use it the exact same way as before, but it
also makes it easy to add new splits or windows if desired. I frequently
want this because I'm launching nexd, and then I want to run additional
commands either to run a client or server or some kind to test
connectivity.

Signed-off-by: Russell Bryant <rbryant@redhat.com>