docs: Update quickstart with docker instructions
For quick testing and demo purposes, provide instructions for
launching a nexd container that can be quickly connected to our demo
service.

Also change the bash history and instructions printed out at startup
when used via these instructions vs. when it's used in development
with `make run-nexd-container`. For the general user case, we don't
need the extra instructions for hitting a local development instance.

Signed-off-by: Russell Bryant <rbryant@redhat.com>