nexd: Don't reconcile with stun when behind symmetric NAT
The code that automatically reconciles the public NAT binding was not
getting skipped if we had already determined we are behind symmetric
NAT. This simple change fixes the flapping observed when behind
symmetric NAT.

There is another bug still present, though. This code should be
updated further to detect moving in and out of networks with symmetric
NAT. I will file a follow-bug on that scenario.

Closes #1001
Closes #1044

Signed-off-by: Russell Bryant <rbryant@redhat.com>